Execute otcamera service after all active jobs are dispatched
The problem is that the same log file is written on every boot
although it should be written to a new file depending on the timestamp.
It could be that after system reboot the otcamera.service is started
before the system time is synced with the RTC causing the system
to return the last saved time before the reboot.

Making the service start after clock synchronization should fix
this problem.